+++ cli/Changelog.md
@@ -1,3 +1,9 @@
+## v2.0.0
+
+- Added "--mods" option, to control the mods that are loaded when running. All other mods are now disabled by default.
+- The default factorio data directory is now `./factorio-test-data-dir` instead of just `./factorio-test-data`.
+- Improved help text.
+
 ## v1.0.5
 
 - Fixed problems when used on Windows.